Resolve #2794: OnlineIndexer/OnlineIndexScrubber: move the sync session lease length to Config
  As a code cleanup, and following PR #2788, move the sync session parameters to OnlineIndexOperationConfig.
